+++ internal/app/machined/pkg/runtime/v1alpha2/v1alpha2_controller.go
@@ -164,7 +164,6 @@ func (ctrl *Controller) Run(ctx context.Context, drainer *runtime.Drainer) error
 		&block.SymlinksController{},
 		&block.SystemDiskController{},
 		&block.UserDiskConfigController{},
-		&block.UserVolumeConfigController{},
 		&block.VolumeConfigController{
 			V1Alpha1Mode: ctrl.v1alpha1Runtime.State().Platform().Mode(),
 			MetaProvider: ctrl.v1alpha1Runtime.State().Machine(),
